First and Last Massive Push
In an effort, that albeit got a little carried away, I set about laying out the foundation for skills.
I now have a large swath of player statistics, skills, and projectile behaviour carved out, enough to have visual simulations of potential projectiles.

Skills are similar to the UML in the GDD, and in short order they will be the same, when I patch the UML to look the same.

Projectiles are a wild card in terms of the GDD, surely now that I realise they are as sophisticated as they are, they deserve a section in the GDD too.

Player statistics is my first proper attempt at using PropertyDrawer, and it is successful, which spurs me on to continue this escapade of frivolous PropertyDrawering (it isn't frivolous, it seriously improves usability of the sometimes hideous Unity inspector serialized representions. This will next be done to ProjectileBehaviour, and finally as far as I can see for now, to Skills, as they have properties that are specific to certain types of skill, and can therefore be hidden in the same fashion in the inspector when not pertinant.

Finally.
I imported so packages, this would be Utility, and PostProcessingStack so we have pretty post processing (they have way prettier on the asset store for mere money, if we get around to it, and decide it is worth the investment, like after we have capital).

And materials, we have some, I made a basic vertex colour supporting surface shader, which I plan to expand to support the previously mentioned (in person to Justin) vertex colour acting as a map against a colour ramp (so we can replace colour ramps, and have entirely new colour configurations on the ship, I have set up a palette for this in MagikaVoxel)

REMINDER
Make sure Justin has MagikaVoxel colour palette I want to go forward with.
Make sure Justin can use the stuff I make.