Fixed the ignore some more.
For whatever reason Git assumes you will only ever have the git ignore in a very specific area of the folder structure of the project, which in this case is assumed to be one layer deeper than we have it, so all the ignores are a little fucky.

This should now resolve that by adding the assumed paths for all ignoreable assets to include one deeper /Scavenger/ layer.