Statistics, Player, and Projectiles
Statistics
Added implicit casting to all supported types (save object, use Get<T>()).

IStatistics
This is the interface responsible for statistics array accessor property and Has() method that both PlayerUEI and ProjectileUEI implement (as well as in the future all other users of statistics).

PlayerUEI
Added array accessor overloading, to access statistics, use [string index], set is safe, will overwrite the value, or add the value as needed, get is not, if an attempt at getting a value that does not exist is attempted it will crash. I should change that.

Projectile
Shifted some logic around.
Projectiles now use Statistics.
Projectile Movement has 2 behaviours so far
ProjectileMovementUEI necessary for projectile movement logic.
ProjectileMovementSequencer acts as a statistic sequencer for the projectile, will have name changed in coming pushes to ProjectileStatisticSequencer as that is more correctly what it is.