Surfaces Star and Title Sequence and Logo Changes
Made Title sequences more responsive, faster, more aesthetically pleasing.
Modernized the logo, in turn no longer need the constellations for the logo.

Added a shader that handles accurate (ish) star colours and allows for multiple textures to be mapped based on temperature of star, in kelvin (this is functioning but has some more work to be done.)