Surfaces.Star Modifications
Increasing the level of realism in the shader all around. Adding a very useful atmospheric scattering like effect to the custom WrapLambert lighting model, as well as a subsurface scattering like effect to simulate the outer layer of the atmosphere. This will need modifications in order to work on solid planets, as this treats the gas giant as being partially transparent in the higher surface regions.

As well I have begun efforts to create a wrapper class for the shader that enables easy and interactable customisation of the shader, instead of having to fiddle around with the shader parameters in the shader itself.
(This can invariably also lead to sequencers)