Stars, Ships, and Sticks
So I still haven't solved the issues I am facing with Input. I have no way of dynamically connecting the input system together yet through the unity inspector, but I do have an idea!
Strings! Just make every input log its last input to a statistic on the player, and pass the string name of that statistic to any listeners, they can then query it, or better yet, they can query it once, and then save a reference to the statistic and just shut up about it afterwards. Whichever.

Stars, found a program that makes pretty skyboxes (as pretty as you make them at least). I decided to put the zip in the git, so we can keep reference to the program. It is free, has no licensing issues, and is lovely (a bit crashy).

Finally, improved the VCMapped materials. So first, they now categorise in the shaders selection menu (this may get more work as I find better ways of naming them). Second, they now use a version of Triplanar mapping, which vastly improves the previously lackluster xy orthographic mapping I was using (which made a whole slew of face alignments look like garbage). I also cleared up some misunderstandings I was having with the inspector for the shader, by first placing a number in front of each property pertaining to what of the 5 materials this property affects, and second, reorder the properties so that all of the non-2D are at the top, and the 2D are at the  bottom (visually this seperates the distinct texture properties better).