Systems, Stars, and Planets 3
Further changes to system generation:
Seperated environment data population from game object population. In this way it is now possible to create an environment, assign it a specific location in the universe and ask it what it would look like without also generating a system.
System game object generation is now tailored entirely to the system view in navigation (previously every bit of the generator was).

Added some new star textures in the hopes of eventually getting a refined star appearance. Current textures are acceptable but more are necessary for higher variability between stars.

Trying to fix an editor freezing bug, I think it is due to the coroutine in Environment but I have yet to discover what exactly is causing it, as the system generation appears to complete, so I don't seem to have any infinite loops, or otherwise invalid logic (still looking though). Observation tells me that it freezes when generating a new system, especially if there is already an existing system, it occurs with frequency, but I have yet to find a pattern. I have a hypothesis that it may still be an infinite loop, even after my efforts to ensure there are none, that only occurs on some hashes, but when I attempt to pause and view the stack all executing code is not from within my project, so I have no idea.

Orbits have been changed drastically. Stars now only show orbits if they are in a binary pair, and only for a single binary pair, any further attempt at predicting where they would be was proving incredibly difficult to determine. Planets also show their orbits, and orbits have been designed to rotate to wherever the star or planet is in the orbit relative to the center, as orbits now fade over their circle (hopefully this makes it harder for people to spot any inaccuracies in the orbits). This is all bugged ever since I changed the system generator to a 2 pass. The planets all have the wrong positions of their stars.

Finally, I created some mock star materials to document some of the basic statistics necessary for various stellar classes.