Systems, Stars, and Planets 2
So I really thought I had already committed a #2 in this sequence, so since I didn't here it is:

Fixed any issues caused by splitting the generation into 2 passes, and optimising. So the bug with incorrect planet positioning was also affecting stars, which I also noticed as it happens.

Stars in binary pairs (such as for now the stars at 0,0) should always be on opposite sides of their little dance, and they weren't which led me to discover that I was incorrectly setting their position (it was being set and offset by the center of mass of the system before the center of mass has been calculated fully).
Planets suffered from a similar issue, which in this case was because their stars positions were not where they should be (due to the above bug).
Everything is now fine.

I removed the pass that checks if planet generation would fail, I mean it means planets can now sometimes be literally inside stars, but I feel like later as I add generation decorators that only gives me more options for conditions to look for and add dynamic new content via. Especially since with the failed planets this system which would generate 7 planets instead generated only 3, because the rest failed to be far enough away from either parent star (it was a pass following the assumption that they would not have formed in that configuration because the star's immense temperature would impede the formation of a star at that distance) I have changed my logic to conclude that systems generate as static pictures of them in their current state, not in an ideal perfectly generated systems state. So now planets can totally generate on course to crashing into a star. (This reminds me I should add logic to the directionality of the orbit so that in those cases planets that generate looking as if they have some how passed through a star instead have their orbit direction flipped, creating the above situation)

Okay. Now I can start to figure out what to do with planets, stars, and galactic generation.