Casestudy ColMon 1 (165863, 15513)
Working on polishing up Planetary collision case study ColMon 1.

Research revealed that the ejecta ring is just that, a ring, or conic section based on the size, velocity, mass, and hardness of both planets, such that in my case I can simplify it to be a ring of ejecta that move away from the event horizon of the collision, towards the smaller of the 2 colliding bodies.

In this case, that means essentially perpendicular to the normal of collision between moon A and moon B.

This means I have created collision assets, such as materials to assign to the particle systems. Particle systems as well have been created, primarily the ejecta ring particle system, and the asteroid collision particle system.
A new shader was created for the asteroid impact explosion, so that it could be the best of 2 worlds, a spherical mesh with faded edges (so it glows in the middle no matter what point of view, and a soft edge particle shader, so that the expanding edge of the explosion fades when close to the planet. As well a particle attractor based on gravity has been constructed, and each of the major bodies have been given one, that would be the Planet, Moon A, and Moon B.

Next steps are to create materials for both colliding bodies that make them appear cracked, glowing, and severly damaged by the impact. As well as to possibly create a behaviour that modifies their spheres.