System Generation
I have implemented a decorator based system generation, where the previously base, and hard coded system generation has been moved to a so called StandardSystem ISystemGenerationDecorator which in its first pass System() generates all the necessary data, in its second pass Star() creates all the stars, in its third pass Planet() creates all the planets and moons, and in its final pass Dungeon() generates all the dungeons.

I also implemented a SingularitySystem  ISystemGenerationDecorator which has a 1 in 1000 chance of converting a system into a singularity system with 1 singularity, it is able to modify the system removing planets and stars as well.

Oh yeah, a bunch of MSE textures for the large pool of world assets, continuing to make the necessary assets for planet generation.

I am pushing because I have a major planned overhaul to the already overhauled system generation. It will be the final overhaul, and it will enable me to have much broader control over the exact nature of any system.
I am going to shift planet temperature calculations away from the system generation passes, where it currently resides, to an element of the EnvironmentBasedPlanet MonoBehaviour, so that dynamic changes to the position of the planet, or position of the star, size of the star, and temperature of the star will have the correct effect on the planet. Essentially I am going to make the systems live simulations instead of static ones. This may make things more expensive to simulate, but I intend on only doing calculations every second, and otherwise leaving them to lerp from the last calculation to the current.