Widgets, Tags, Dungeons, and Systems
Widgets
Added a match position widget behaviour that makes 2 connected widgets that are in seperate parts of the UI hierarchy move together (so that a label can be infront of other stuff but still move with the thing it is infront of).
WidgetSignalTag is a behaviour that given a tag looks it up in the scheme and assigns its label and icon to match, the icon also matches scheme colouring. This is used to show what tags are attached to dungeons both in the system navigation screen, and in the dungeon description widget.
WidgetSignalTagGrid is the behaviour responsible for populating the dungeon description with the dungeon tags.
Added UIScreens to better sort layered content in the UI, as well as facilitating fast and easy transitioning between disparate sections of UI content.
WidgetWorldPosition matches the position of a RectTransform to the position of a Transform, keeping it on screen, and either preventing overlap as close to exactly as possible, or by simply pushing all WidgetWorldPositions away from eachother.
WidgetSignalDescription is the widget behaviour that handles the description aspect of each dungeon.
WidgetSignal is the dungeon selector on the system navigation screen, it shows the name of the dungeon on hover, as well as the most important tags, and matches the position of the dungeon in world space in UI space (using WidgetWorldPosition).
WidgetScheme handles all the colour and symbol information for the UI.
UIManager handles the UIScreens, switching between them, setting their active states accordingly, and adding and removing widgets from the screens.

Added tag symbols, made in inkscape.
Have tags for:
Allied
Neutral
Hostile
Defended
Buy
Sell
Respawn
Ship
Ships
Debris
Asteroid
Asteroids
Station (and all variants)
Asteroid Station (and all variants)
Shipwreck
Graveyard
Conflict
Battle
Repair
Renamed a bunch of the old ones using the whole game_nav_signal_***.png

Handled environment better, in preparation for the systems to system transitioning, using Environment a static class that holds information between scenes.
EnvironmentSetter handles individual monobehaviours that want to set the static variables, for inspector interfacing with purely static variables.

Systems has been started, this is the scene that shows all the systems around the player and allows them to select one to transition to a new system.