Significant Progress To Systems -> System
Significantly improving the Systems screen to better utilize existing details to express more context for the player. I am going to add the ship to the Systems map as soon as I can, and I am going to start offsetting the stars up and down a little bit, so it looks more spacey, and then when a new system is selected the ship moves there.

Okay so what has been done:
Attempted to add a cursor and edge panning to the Systems camera, both are non functional right now.

Updated dungeon loader to utilise Resources, now moddable.
Expanded DungeonType to represent new static 'definitions' loaded when the dungeon definitions are, from dungeons, via Resources.
These are: categories, target, and all tags.
Added more parameters to Environment, to represent changes made to the system generation and systems needs.
These are: environment time information, dungeon limits (for passives to alter) jump fuel and range parameters, selected coordinates, and generate and populate booleans, as well as the scene name to load in the load scene, and the async loading operation for the currently loading scene.
Updated EnvironmentSetter to represent the new Environment parameters.
Updated EnvironmentTime to represent the new Environment parameters.
Updated ISystemGeneratorDecorator and SystemGeneratorDecorator to represent the new definitions for all system generation passes.
Attempting to switch system generation over to a coroutinable system that can take as long as needed to load without causing lag.
Updated System decorators to represent the changes in the ISystemGeneratorDecorator.
Moved some basic system environment statistics to the new System pass of the System Generator Decorators, allowing the Systems map to determine where and what dungeons will be in each system, to offer more context to the system data.
Updated SystemGenerator to better utilise the new Coroutine format for system generation (currently primarily non utilised).
Added a LoadingBar behaviour, which doesn't work for my scenes because they generate post loading, and so the 'load' occurs almost instantly, but it still lags right at the frame of scene switch (as the new system / systems scene generates a significant quantity of data and game objects).
Added a rect transform animator using a newer format to allow better adjustment for the exacting needs of each UI animation circumstance.
Added and split up all the Dungeon definition JSON files, these are any, asteroid, asteroids, planet, star, these are seperated into different files per target (this was my decision, and not a requirement).
Added a SystemsCamera, it allows zooming in and out, and soon will implement an offset system to better utilise the upcoming changes to the Systems (primarily the added z offset for each star system).
SystemSceneSwitch which handles switching to System from a single system in Systems.
SystemsFilter is a container that enables UI to determine what data the UI should most focus on in efforts to contextualize it.
SystemsGenerator generates a grid of systems within the players jump range, allowing for the player to view all systems they could potentially reach if they have enough fuel.
SystemsSceneSwitch which handles loading the Systems scene from the System scene, via the back button.
Lots of Icons and logos for all the tags and dungoens to the widget scheme.
WidgetScheme now propegates through scenes. Should add support for naming schemes, and allowing ignoring, defining, updating, or referring, which would allow schemes to override eachother, ignore eachother, update eachother, or refer to a preexisting named definition.
Updated widget signal to support all the new labels.
Fixed WidgetWorldPosition to prevent it from clamping position if no screen edge is defined.
Added a ListHelper for shuffle functionality.
Added new parse functions to StringHelper.
Added a widget for the date
Added a widget for the fuel
Added a widget for searching
Added a widget for each system in Systems
Added a widget for each tag on each system in Systems