Navigation Merger
Systems and System scenes have been merged, and Systems navigation has been overhauled.

Added support for literals and language libraries.

Switch all if not almost all widgets and other context conveyers to use the new literals system, given dungeon information is no longer the raw string to display, but instead to use that string as a key in the dictionary of the current language.