Options and Resources
Added Sprites which handles loading sprites from resources for schemes and others.

Added some float extensions (RoundTo which rounds to the required integer number of decimal places)

Options have value sand definitions, not sure where I am going with this, ultimately may want to merge the OptionValue with Statistic, and with Statistic from ships. They all represent the same problem and solution, all done seperately which smacks of unification.

Moved files around for the new resources and sprite loading.

Added some more ships.

Broke InventoryShips. I need to rebuild it with the new requirements, that it be able to display all ships, or only unlocked ships. Currently it is just a buggy pile of broken.